Fixed error view add dailylogs/images
Removed the callback function and handle set for downloading images, was
causing an error when adding or viewing daily logs

Aded copying image files to the dropbox directory no matter the location
of the original file.
Also added a number format string to force correct date format for file
names of daily logs.