Updated location to look for new version
Since the version number does not exist on the main github page, it now
searches for the current version in the repository.  Current versions of
YCWeather need to be updated manually to this version to ensure it can
automatically contact github for an automatic update.

Also, installer_script.iss updated to match local directories of
djwalters for compiling and releasing new versions.